requirements.txt

python-telegram-bot
requests
python-dotenv
